Module paperclip::actix

Plugin types, traits and macros for actix-web framework.

Modules

web

Proxy module for actix_web::web.

Structs

AcceptedJson

App

Wrapper for actix_web::App.

CreatedJson

HttpResponseWrapper

Workaround for possibility to directly return HttpResponse from closure handler.

NoContent

ResponderWrapper

Wrapper for wrapping over impl Responder thingies (to avoid breakage).

ResponseWrapper

Traits

Mountable

Indicates that this thingmabob has a path and a bunch of definitions and operations.

OpenApiExt

Extension trait for actix-web applications.

OperationModifier

Actix-specific trait for indicating that this entity can modify an operation and/or update the global map of definitions.

Attribute Macros

api_v2_errors

Marker attribute for indicating that the marked object can represent non-2xx (error) status codes with optional descriptions.

api_v2_errors_overlay

Marker attribute for indicating that the marked object can filter error responses from the the #[api_v2_errors] macro.

api_v2_operation

Marker attribute for indicating that a function is an OpenAPI v2 compatible operation.

delete

Creates route handler with paperclip::actix::web::Resource. In order to control the output type and status codes the return value/response must implement the trait actix_web::Responder.

get

Creates route handler with paperclip::actix::web::Resource. In order to control the output type and status codes the return value/response must implement the trait actix_web::Responder.

patch

Creates route handler with paperclip::actix::web::Resource. In order to control the output type and status codes the return value/response must implement the trait actix_web::Responder.

post

Creates route handler with paperclip::actix::web::Resource. In order to control the output type and status codes the return value/response must implement the trait actix_web::Responder.

put

Creates route handler with paperclip::actix::web::Resource. In order to control the output type and status codes the return value/response must implement the trait actix_web::Responder.

Derive Macros

Apiv2Header

Derive attribute for indicating that a type is an OpenAPI v2 compatible header parameter.

Apiv2Schema

Derive attribute for indicating that a type is an OpenAPI v2 compatible definition.

Apiv2Security

Marker attribute for indicating that an object forbids public access to operation (for example AccessToken).