Struct paperclip::actix::App

pub struct App<T, B> { /* fields omitted */ }

Wrapper for actix_web::App.

Implementations

impl<T, B> App<T, B> where
B: MessageBody,
T: ServiceFactory<Config = (), Request = ServiceRequest, Response = ServiceResponse<B>, Error = Error, InitError = ()>,

pub fn data<U>(self, data: U) -> App<T, B> where
U: 'static,

Proxy for actix_web::App::data.

NOTE: This doesn’t affect spec generation.

pub fn data_factory<F, Out, D, E>(self, data: F) -> App<T, B> where
F: 'static + Fn() -> Out,
Out: 'static + Future<Output = Result<D, E>>,
D: 'static,
E: Debug,

Proxy for actix_web::App::data_factory.

NOTE: This doesn’t affect spec generation.

pub fn app_data<U>(self, data: U) -> App<T, B> where
U: 'static,

Proxy for actix_web::App::app_data.

NOTE: This doesn’t affect spec generation.

pub fn configure<F>(self, f: F) -> App<T, B> where
F: FnOnce(&mut ServiceConfig<'_>),

Wrapper for actix_web::App::configure.

pub fn route(self, path: &str, route: Route) -> App<T, B>

Wrapper for actix_web::App::route.

pub fn service<F>(self, factory: F) -> App<T, B> where
F: 'static + Mountable + HttpServiceFactory,

Wrapper for actix_web::App::service.

pub fn default_service<F, U>(self, f: F) -> App<T, B> where
F: IntoServiceFactory<U>,
U: 'static + ServiceFactory<Config = (), Request = ServiceRequest, Response = ServiceResponse<Body>, Error = Error, InitError = ()>,
<U as ServiceFactory>::InitError: Debug,

Proxy for actix_web::App::default_service.

NOTE: This doesn’t affect spec generation.

pub fn external_resource<N, U>(self, name: N, url: U) -> App<T, B> where
N: AsRef<str>,
U: AsRef<str>,

Proxy for actix_web::App::external_resource.

NOTE: This doesn’t affect spec generation.

pub fn wrap<M, B1>(
self,
mw: M
) -> App<impl ServiceFactory<Config = (), Request = ServiceRequest, Response = ServiceResponse<B1>, Error = Error, InitError = ()>, B1> where
M: Transform<<T as ServiceFactory>::Service, Request = ServiceRequest, Response = ServiceResponse<B1>, Error = Error, InitError = ()>,
B1: MessageBody,

Proxy for actix_web::web::App::wrap.

NOTE: This doesn’t affect spec generation.

pub fn wrap_fn<B1, F, R>(
self,
mw: F
) -> App<impl ServiceFactory<Config = (), Request = ServiceRequest, Response = ServiceResponse<B1>, Error = Error, InitError = ()>, B1> where
B1: MessageBody,
F: FnMut(ServiceRequest, &mut <T as ServiceFactory>::Service) -> R + Clone,
R: Future<Output = Result<ServiceResponse<B1>, Error>>,

Proxy for actix_web::web::App::wrap_fn.

NOTE: This doesn’t affect spec generation.

pub fn with_json_spec_at(self, path: &str) -> App<T, B>

Mounts the specification for all operations and definitions recorded by the wrapper and serves them in the given path as a JSON.

pub fn with_json_spec_v3_at(self, path: &str) -> App<T, B>

Converts the generated v2 specification to v3 and then mounts the v3 specification for all operations and definitions recorded by the wrapper and serves them in the given path as a JSON.

pub fn with_raw_json_spec<F>(self, call: F) -> App<T, B> where
F: FnMut(App<T, B>, Value) -> App<T, B>,

Calls the given function with App and JSON Value representing your API specification built until now.

NOTE: Unlike with_json_spec_at, this only has the API spec built until this function call. Any route handler added after this call won’t affect the spec. So, it’s important to call this function after adding all route handlers.

pub fn with_raw_json_spec_v3<F>(self, call: F) -> App<T, B> where
F: FnMut(App<T, B>, Value) -> App<T, B>,

Calls the given function with App and JSON Value representing your API v2 specification built until now which is converted to v3.

NOTE: Unlike with_json_spec_at, this only has the API spec built until this function call. Any route handler added after this call won’t affect the spec. So, it’s important to call this function after adding all route handlers.

pub fn with_swagger_ui_at(self, path: &str) -> App<T, B>

Exposes the previously built JSON specification with Swagger UI at the given path

NOTE: you MUST call with_json_spec_at before calling this function

pub fn build(self) -> App<T, B>

Builds and returns the actix_web::App.

pub fn trim_base_path(self) -> App<T, B>

Trim’s the Api base path from the start of all method paths. NOTE: much like with_raw_json_spec this only has the API spec built until this function call. Any route handler added after this call won’t have the base path trimmed. So, it’s important to call this function after adding all route handlers.