Trait paperclip::actix::OpenApiExt

pub trait OpenApiExt<T> {
    type Wrapper;

    fn wrap_api(self) -> Self::Wrapper;
    fn wrap_api_with_spec(
        self, 
        spec: Api<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>, DefaultSchemaRaw>
    ) -> Self::Wrapper;
}

Extension trait for actix-web applications.

Required Associated Types

type Wrapper

Required Methods

fn wrap_api(self) -> Self::Wrapper

Consumes this app and produces its wrapper to start tracking paths and their corresponding operations.

fn wrap_api_with_spec(
self,
spec: Api<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>, DefaultSchemaRaw>
) -> Self::Wrapper

Same as wrap_api initializing with provided specification defaults. Useful for defining Api properties outside of definitions and paths.