Trait paperclip::actix::OperationModifier

pub trait OperationModifier: Apiv2Schema {
    pub fn update_parameter(
        _op: &mut Operation<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>>
    ) { ... }

    pub fn update_response(
        _op: &mut Operation<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>>
    ) { ... }

    pub fn update_definitions(map: &mut BTreeMap<String, DefaultSchemaRaw>) { ... }

    pub fn update_security(
        op: &mut Operation<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>>
    ) { ... }

    pub fn update_security_definitions(
        map: &mut BTreeMap<String, SecurityScheme>
    ) { ... }
}

Actix-specific trait for indicating that this entity can modify an operation and/or update the global map of definitions.

Provided methods

pub fn update_parameter(
_op: &mut Operation<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>>
)

Update the parameters list in the given operation (if needed).

pub fn update_response(
_op: &mut Operation<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>>
)

Update the responses map in the given operation (if needed).

pub fn update_definitions(map: &mut BTreeMap<String, DefaultSchemaRaw>)

Update the definitions map (if needed).

pub fn update_security(
op: &mut Operation<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>>
)

Update the security map in the given operation (if needed).

pub fn update_security_definitions(map: &mut BTreeMap<String, SecurityScheme>)

Update the security definition map (if needed).

Loading content...