Struct paperclip::actix::web::Form

pub struct Form<T>(pub T);

URL encoded payload extractor and responder.

Form has two uses: URL encoded responses, and extracting typed data from URL request payloads.

Extractor

To extract typed data from a request body, the inner type T must implement the DeserializeOwned trait.

Use FormConfig to configure extraction options.

Examples

use actix_web::{post, web};
use serde::Deserialize;

#[derive(Deserialize)]
struct Info {
    name: String,
}

// This handler is only called if:
// - request headers declare the content type as `application/x-www-form-urlencoded`
// - request payload deserializes into an `Info` struct from the URL encoded format
#[post("/")]
async fn index(web::Form(form): web::Form<Info>) -> String {
    format!("Welcome {}!", form.name)
}

Responder

The Form type also allows you to create URL encoded responses by returning a value of type Form<T> where T is the type to be URL encoded, as long as T implements Serialize.

Examples

use actix_web::{get, web};
use serde::Serialize;

#[derive(Serialize)]
struct SomeForm {
    name: String,
    age: u8
}

// Response will have:
// - status: 200 OK
// - header: `Content-Type: application/x-www-form-urlencoded`
// - body: `name=actix&age=123`
#[get("/")]
async fn index() -> web::Form<SomeForm> {
    web::Form(SomeForm {
        name: "actix".to_owned(),
        age: 123
    })
}

Panics

URL encoded forms consist of unordered key=value pairs, therefore they cannot be decoded into any type which depends upon data ordering (eg. tuples). Trying to do so will result in a panic.

Tuple Fields
0: T
Implementations

impl<T> Form<T>

pub fn into_inner(self) -> T

Unwrap into inner T value.