Struct paperclip::actix::web::HttpResponse

pub struct HttpResponse<B = Body> { /* fields omitted */ }

An HTTP Response

Implementations

impl Response<Body>

pub fn Continue() -> ResponseBuilder

pub fn SwitchingProtocols() -> ResponseBuilder

pub fn Processing() -> ResponseBuilder

pub fn Ok() -> ResponseBuilder

pub fn Created() -> ResponseBuilder

pub fn Accepted() -> ResponseBuilder

pub fn NonAuthoritativeInformation() -> ResponseBuilder

pub fn NoContent() -> ResponseBuilder

pub fn ResetContent() -> ResponseBuilder

pub fn PartialContent() -> ResponseBuilder

pub fn MultiStatus() -> ResponseBuilder

pub fn AlreadyReported() -> ResponseBuilder

pub fn MultipleChoices() -> ResponseBuilder

pub fn MovedPermanently() -> ResponseBuilder

pub fn Found() -> ResponseBuilder

pub fn SeeOther() -> ResponseBuilder

pub fn NotModified() -> ResponseBuilder

pub fn UseProxy() -> ResponseBuilder

pub fn TemporaryRedirect() -> ResponseBuilder

pub fn PermanentRedirect() -> ResponseBuilder

pub fn BadRequest() -> ResponseBuilder

pub fn NotFound() -> ResponseBuilder

pub fn Unauthorized() -> ResponseBuilder

pub fn PaymentRequired() -> ResponseBuilder

pub fn Forbidden() -> ResponseBuilder

pub fn MethodNotAllowed() -> ResponseBuilder

pub fn NotAcceptable() -> ResponseBuilder

pub fn ProxyAuthenticationRequired() -> ResponseBuilder

pub fn RequestTimeout() -> ResponseBuilder

pub fn Conflict() -> ResponseBuilder

pub fn Gone() -> ResponseBuilder

pub fn LengthRequired() -> ResponseBuilder

pub fn PreconditionFailed() -> ResponseBuilder

pub fn PreconditionRequired() -> ResponseBuilder

pub fn PayloadTooLarge() -> ResponseBuilder

pub fn UriTooLong() -> ResponseBuilder

pub fn UnsupportedMediaType() -> ResponseBuilder

pub fn RangeNotSatisfiable() -> ResponseBuilder

pub fn ExpectationFailed() -> ResponseBuilder

pub fn UnprocessableEntity() -> ResponseBuilder

pub fn TooManyRequests() -> ResponseBuilder

pub fn InternalServerError() -> ResponseBuilder

pub fn NotImplemented() -> ResponseBuilder

pub fn BadGateway() -> ResponseBuilder

pub fn ServiceUnavailable() -> ResponseBuilder

pub fn GatewayTimeout() -> ResponseBuilder

pub fn VersionNotSupported() -> ResponseBuilder

pub fn VariantAlsoNegotiates() -> ResponseBuilder

pub fn InsufficientStorage() -> ResponseBuilder

pub fn LoopDetected() -> ResponseBuilder

impl Response<Body>

pub fn build(status: StatusCode) -> ResponseBuilder

Create http response builder with specific status.

pub fn build_from<T>(source: T) -> ResponseBuilder where
T: Into<ResponseBuilder>,

Create http response builder

pub fn new(status: StatusCode) -> Response<Body>

Constructs a response

pub fn from_error(error: Error) -> Response<Body>

Constructs an error response

pub fn into_body<B>(self) -> Response<B>

Convert response to response with body

impl<B> Response<B>

pub fn with_body(status: StatusCode, body: B) -> Response<B>

Constructs a response with body

pub fn head(&self) -> &ResponseHead

Http message part of the response

pub fn head_mut(&mut self) -> &mut ResponseHead

Mutable reference to a http message part of the response

pub fn error(&self) -> Option<&Error>

The source error for this response

pub fn status(&self) -> StatusCode

Get the response status code

pub fn status_mut(&mut self) -> &mut StatusCode

Set the StatusCode for this response

pub fn headers(&self) -> &HeaderMap

Get the headers from the response

pub fn headers_mut(&mut self) -> &mut HeaderMap

Get a mutable reference to the headers

pub fn cookies(&self) -> CookieIter<'_>

Get an iterator for the cookies set by this response

pub fn add_cookie(&mut self, cookie: &Cookie<'_>) -> Result<(), Error>

Add a cookie to this response

pub fn del_cookie(&mut self, name: &str) -> usize

Remove all cookies with the given name from this response. Returns the number of cookies removed.

pub fn upgrade(&self) -> bool

Connection upgrade status

pub fn keep_alive(&self) -> bool

Keep-alive status for this connection

pub fn extensions(&self) -> Ref<'_, Extensions>

Responses extensions

pub fn extensions_mut(&mut self) -> RefMut<'_, Extensions>

Mutable reference to a the response’s extensions

pub fn body(&self) -> &ResponseBody<B>

Get body of this response

pub fn set_body<B2>(self, body: B2) -> Response<B2>

Set a body

pub fn into_parts(self) -> (Response<()>, ResponseBody<B>)

Split response and body

pub fn drop_body(self) -> Response<()>

Drop request’s body

pub fn map_body<F, B2>(self, f: F) -> Response<B2> where
F: FnOnce(&mut ResponseHead, ResponseBody<B>) -> ResponseBody<B2>,

Set a body and return previous body value

pub fn take_body(&mut self) -> ResponseBody<B>

Extract response body