Struct paperclip::v2::models::DefaultSchemaRaw

pub struct DefaultSchemaRaw {
    pub reference: Option<String>,
    pub title: Option<String>,
    pub description: Option<String>,
    pub data_type: Option<DataType>,
    pub format: Option<DataTypeFormat>,
    pub example: Option<Value>,
    pub properties: BTreeMap<String, Box<DefaultSchemaRaw, Global>, Global>,
    pub items: Option<Box<DefaultSchemaRaw, Global>>,
    pub enum_: Vec<Value, Global>,
    pub extra_props: Option<Either<bool, Box<DefaultSchemaRaw, Global>>>,
    pub required: BTreeSet<String, Global>,
    /* private fields */
}

Raw version of schema.

NOTE: This doesn’t have smart pointers to reuse definitions throughout the spec. Instead, it contains the actual schema with unresolvable $ref fields.

Default schema if your schema doesn’t have any custom fields.

https://github.com/OAI/OpenAPI-Specification/blob/master/versions/2.0.md#schemaObject

Fields
reference: Option<String>title: Option<String>description: Option<String>data_type: Option<DataType>format: Option<DataTypeFormat>example: Option<Value>properties: BTreeMap<String, Box<DefaultSchemaRaw, Global>, Global>items: Option<Box<DefaultSchemaRaw, Global>>enum_: Vec<Value, Global>extra_props: Option<Either<bool, Box<DefaultSchemaRaw, Global>>>required: BTreeSet<String, Global>
Implementations

impl DefaultSchemaRaw

pub fn remove_refs(&mut self)

Recursively removes all $ref values in this schema.

pub fn retain_ref(&mut self)

Recursively removes all properties other than $ref value if the $ref is non-null.