paperclip::v2::models
Type Alias ResolvableApi

pub type ResolvableApi<S> = Api<Arc<RwLock<Parameter<Resolvable<S>>>>, Arc<RwLock<Response<Resolvable<S>>>>, Resolvable<S>>;

OpenAPI v2 spec which can be traversed and resolved for codegen.

Aliased Type

struct ResolvableApi<S> {

    pub swagger: Version,
    pub definitions: BTreeMap<String, Resolvable<S>>,
    pub paths: BTreeMap<String, PathItem<Arc<RwLock<Parameter<Resolvable<S>>>>, Arc<RwLock<Response<Resolvable<S>>>>>>,
    pub host: Option<String>,
    pub base_path: Option<String>,
    pub consumes: BTreeSet<MediaRange>,
    pub produces: BTreeSet<MediaRange>,
    pub schemes: BTreeSet<OperationProtocol>,
    pub parameters: BTreeMap<String, Arc<RwLock<Parameter<Resolvable<S>>>>>,
    pub responses: BTreeMap<String, Arc<RwLock<Response<Resolvable<S>>>>>,
    pub security_definitions: BTreeMap<String, SecurityScheme>,
    pub security: Vec<BTreeMap<String, BTreeSet<String>>>,
    pub tags: Vec<Tag>,
    pub external_docs: Option<ExternalDocs>,
    pub coders: Coders,
    pub support_crates: BTreeMap<String, String>,
    pub spec_format: SpecFormat,
    pub info: Info,
    pub extensions: BTreeMap<String, Value>,

}

Fields
swagger: Versiondefinitions: BTreeMap<String, Resolvable<S>>paths: BTreeMap<String, PathItem<Arc<RwLock<Parameter<Resolvable<S>>>>, Arc<RwLock<Response<Resolvable<S>>>>>>host: Option<String>base_path: Option<String>consumes: BTreeSet<MediaRange>produces: BTreeSet<MediaRange>schemes: BTreeSet<OperationProtocol>parameters: BTreeMap<String, Arc<RwLock<Parameter<Resolvable<S>>>>>responses: BTreeMap<String, Arc<RwLock<Response<Resolvable<S>>>>>security_definitions: BTreeMap<String, SecurityScheme>security: Vec<BTreeMap<String, BTreeSet<String>>>tags: Vec<Tag>external_docs: Option<ExternalDocs>coders: Coders

Extension for custom coders to be used for decoding API objects.

An example for JSON would be:

x-rust-coders:
  application/json:
    encoder_path: serde_json::to_writer
    decoder_path: serde_json::from_reader
    any_value: serde_json::Value
    error_path: serde_json::Error

NOTE: JSON and YAML encodings are already supported.

support_crates: BTreeMap<String, String>

Additional crates that need to be added to the manifest.

The key is the LHS of a dependency, which is the crate name. The value is the RHS of a crate’s requirements as it would appear in the manifest. Note that the caller must add proper escaping wherever required.

For example, the following are all valid:

my_crate: "0.7"

my_crate: "{ git = \"git://foo.bar/repo\" }"

my_crate: "{ version = \"0.9\", features = [\"booya\"] }"

spec_format: SpecFormat

This field is set manually, because we don’t know the format in which the spec was provided and we need to use this as the fallback encoding.

info: Infoextensions: BTreeMap<String, Value>