Trait paperclip::v2::schema::Apiv2Schema

pub trait Apiv2Schema {
    fn name() -> Option<String> { ... }
    fn description() -> &'static str { ... }
    fn required() -> bool { ... }
    fn raw_schema() -> DefaultSchemaRaw { ... }
    fn schema_with_ref() -> DefaultSchemaRaw { ... }
    fn security_scheme() -> Option<SecurityScheme> { ... }
    fn header_parameter_schema() -> Vec<Parameter<DefaultSchemaRaw>, Global>  { ... }
}

Represents a OpenAPI v2 schema object convertible. This is auto-implemented by framework-specific macros:

Apiv2Schema for schema objects.

Apiv2Security for security scheme objects.

Apiv2Header for header parameters objects.

This is implemented for primitive types by default.

Provided Methods

fn name() -> Option<String>

Name of this schema. This is the name to which the definition of the object is mapped.

fn description() -> &'static str

Description of this schema. In case the trait is derived, uses the documentation on the type.

fn required() -> bool

Indicates the requirement of this schema.

fn raw_schema() -> DefaultSchemaRaw

Returns the raw schema for this object.

fn schema_with_ref() -> DefaultSchemaRaw

Returns the schema with a reference (if this is an object).

Here, we set the global reference to this object using its name, and we either remove the reference (remove_refs) or remove all properties other than the reference (retain_ref) based on where we’re storing this object in the spec i.e., in an operation/response or in the map of definitions.

And we do that because at the time of this writing, statically collecting all models for all operations involved a lot of work, and so I went for runtime collection. Even though this happens at runtime, we’re only doing this once at the start of the application, so it won’t affect the incoming requests at all.

fn security_scheme() -> Option<SecurityScheme>

Returns the security scheme for this object.

fn header_parameter_schema() -> Vec<Parameter<DefaultSchemaRaw>, Global>