Struct paperclip::v2::EmitterState

pub struct EmitterState {
    pub working_dir: PathBuf,
    pub ns_sep: &'static str,
    pub mod_prefix: &'static str,
    /* private fields */
}

Holds the state for your schema emitter.

Fields
working_dir: PathBuf

Working directory - the path in which the necessary modules are generated.

ns_sep: &'static str

Namespace separation string.

mod_prefix: &'static str

Module prefix for using in generated code.

Implementations

impl EmitterState

pub fn set_url(&self, url: &str) -> Result<(), Error>

Sets the base URL for this session.

NOTE: Once Emitter::generate is called, this gets overridden by host and basePath fields in spec (if they exist).

pub fn base_url(&self) -> String

Base URL for this API.

NOTE: Once Emitter::generate is called, this gets overridden by host and basePath fields in spec (if they exist).

impl EmitterState

pub fn set_meta(&self, meta: CrateMeta)

Sets the crate metadata for this session.