Crate paperclip

Paperclip is a OpenAPI code generator for efficient type-safe compile-time checked HTTP APIs in Rust.

See the website for detailed documentation and examples.

Modules

actix
Plugin types, traits and macros for actix-web framework.

util

v2
Utilities related to the OpenAPI v2 specification.

Enums

PaperClipError
Global error which encapsulates all related errors.

Type Definitions

PaperClipResult
Generic result used throughout this library.

Attribute Macros

api_v2_schema
Converts your struct to support deserializing from an OpenAPI v2 Schema object (example). This adds the necessary fields (in addition to your own fields) and implements the Schema trait for parsing and codegen.