Trait paperclip::actix::Mountable

pub trait Mountable {
    fn path(&self) -> &str;

    fn operations(
        &mut self
    ) -> BTreeMap<HttpMethod, Operation<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>>>;

    fn definitions(&mut self) -> BTreeMap<String, DefaultSchemaRaw>;

    fn security_definitions(&mut self) -> BTreeMap<String, SecurityScheme>;

    fn update_operations(
        &mut self, 
        map: &mut BTreeMap<String, PathItem<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>>>
    ) { ... }
}

Indicates that this thingmabob has a path and a bunch of definitions and operations.

Required methods

fn path(&self) -> &str

Where this thing gets mounted.

fn operations(
&mut self
) -> BTreeMap<HttpMethod, Operation<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>>>

Map of HTTP methods and the associated API operations.

fn definitions(&mut self) -> BTreeMap<String, DefaultSchemaRaw>

The definitions recorded by this object.

fn security_definitions(&mut self) -> BTreeMap<String, SecurityScheme>

The security definitions recorded by this object.

Provided methods

fn update_operations(
&mut self,
map: &mut BTreeMap<String, PathItem<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>>>
)

Updates the given map of operations with operations tracked by this object.

NOTE: Overriding implementations must ensure that the PathItem is normalized before updating the input map.