Trait paperclip::actix::Mountable

pub trait Mountable {
    pub fn path(&self) -> &str;

    pub fn operations(
        &mut self
    ) -> BTreeMap<HttpMethod, Operation<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>>>;

    pub fn definitions(&mut self) -> BTreeMap<String, DefaultSchemaRaw>;

    pub fn security_definitions(&mut self) -> BTreeMap<String, SecurityScheme>;

    pub fn update_operations(
        &mut self, 
        map: &mut BTreeMap<String, PathItem<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>>>
    ) { ... }
}

Indicates that this thingmabob has a path and a bunch of definitions and operations.

Required methods

pub fn path(&self) -> &str

Where this thing gets mounted.

pub fn operations(
&mut self
) -> BTreeMap<HttpMethod, Operation<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>>>

Map of HTTP methods and the associated API operations.

pub fn definitions(&mut self) -> BTreeMap<String, DefaultSchemaRaw>

The definitions recorded by this object.

pub fn security_definitions(&mut self) -> BTreeMap<String, SecurityScheme>

The security definitions recorded by this object.

Loading content...
Provided methods

pub fn update_operations(
&mut self,
map: &mut BTreeMap<String, PathItem<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>>>
)

Updates the given map of operations with operations tracked by this object.

NOTE: Overriding implementations must ensure that the PathItem is normalized before updating the input map.

Loading content...