Trait paperclip::actix::OperationModifier

pub trait OperationModifier: Apiv2Schema + Sized {
    // Provided methods
    fn update_parameter(
        op: &mut Operation<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>>
    ) { ... }
    fn update_response(
        _op: &mut Operation<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>>
    ) { ... }
    fn update_definitions(map: &mut BTreeMap<String, DefaultSchemaRaw, Global>) { ... }
    fn update_security(
        op: &mut Operation<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>>
    ) { ... }
    fn update_security_definitions(
        map: &mut BTreeMap<String, SecurityScheme, Global>
    ) { ... }
}

Actix-specific trait for indicating that this entity can modify an operation and/or update the global map of definitions.

Provided Methods

fn update_parameter( op: &mut Operation<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>> )

Update the parameters list in the given operation (if needed).

fn update_response( _op: &mut Operation<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>> )

Update the responses map in the given operation (if needed).

fn update_definitions(map: &mut BTreeMap<String, DefaultSchemaRaw, Global>)

Update the definitions map (if needed).

fn update_security( op: &mut Operation<Parameter<DefaultSchemaRaw>, Response<DefaultSchemaRaw>> )

Update the security map in the given operation (if needed).

fn update_security_definitions( map: &mut BTreeMap<String, SecurityScheme, Global> )

Update the security definition map (if needed).