paperclip::actix
Module web

Proxy module for actix_web::web.

Structs

Bytes

A cheaply cloneable and sliceable chunk of contiguous memory.

BytesMut

A unique reference to a contiguous slice of memory.

Data

Application data wrapper and extractor.

Form

URL encoded payload extractor and responder.

FormConfig

Form extractor configuration.

HttpRequest

An incoming request.

HttpResponse

An outgoing response.

Json

JSON extractor and responder.

JsonConfig

Json extractor configuration.

Path

Extract typed data from request path segments.

PathConfig

Path extractor configuration

Payload

Extract a request’s raw payload stream.

PayloadConfig

Configuration for request payloads.

Query

Extract typed information from the request’s query.

QueryConfig

Query extractor configuration.

ReqData

Request-local data extractor.

Resource

Wrapper for actix_web::Resource

Route

Wrapper for actix_web::Route

Scope

Wrapper for actix_web::Scope

ServiceConfig

Wrapper for actix_web::web::ServiceConfig.

Functions

block

Executes blocking function on a thread pool, returns future that resolves to result of the function execution.

delete

Wrapper for actix_web::web::delete.

get

Wrapper for actix_web::web::get.

head

Wrapper for actix_web::web::head.

method

Wrapper for actix_web::web::method.

options

Wrapper for actix_web::web::options.

patch

Wrapper for actix_web::web::patch.

post

Wrapper for actix_web::web::post.

put

Wrapper for actix_web::web::put.

resource

Wrapper for actix_web::web::resource.

scope

Wrapper for actix_web::web::scope.

service

Creates a raw service for a specific path.

to

Creates a new any-method route with handler.