Module paperclip::actix::web

Proxy module for actix_web::web.

Structs

Bytes
A reference counted contiguous slice of memory.

BytesMut
A unique reference to a contiguous slice of memory.

Data
Application data.

Form
Form data helper (application/x-www-form-urlencoded)

FormConfig
Form extractor configuration

HttpRequest
An HTTP Request

HttpResponse
An HTTP Response

Json
Json helper

JsonConfig
Json extractor configuration

Path
Extract typed information from the request’s path.

PathConfig
Path extractor configuration

Payload
Payload extractor returns request ’s payload stream.

PayloadConfig
Configuration for request’s payload.

Query
Extract typed information from the request’s query.

QueryConfig
Query extractor configuration

ReqData
Request-local data extractor.

Resource
Wrapper for actix_web::Resource

Route
Wrapper for actix_web::Route

Scope
Wrapper for actix_web::Scope

ServiceConfig
Wrapper for actix_web::web::ServiceConfig.

Functions

block
Execute blocking function on a thread pool, returns future that resolves to result of the function execution.

delete
Wrapper for actix_web::web::delete.

get
Wrapper for actix_web::web::get.

head
Wrapper for actix_web::web::head.

method
Wrapper for actix_web::web::method.

options
Wrapper for actix_web::web::options.

patch
Wrapper for actix_web::web::patch.

post
Wrapper for actix_web::web::post.

put
Wrapper for actix_web::web::put.

resource
Wrapper for actix_web::web::resource.

scope
Wrapper for actix_web::web::scope.

service
Create raw service for a specific path.

to
Create a new route and add handler.