Struct paperclip::actix::web::Query

pub struct Query<T>(pub T);

Extract typed information from the request’s query.

Note: A query string consists of unordered key=value pairs, therefore it cannot be decoded into any type which depends upon data ordering e.g. tuples or tuple-structs. Attempts to do so will fail at runtime.

QueryConfig allows to configure extraction process.

Example

use actix_web::{web, App};
use serde_derive::Deserialize;

#[derive(Debug, Deserialize)]
pub enum ResponseType {
   Token,
   Code
}

#[derive(Deserialize)]
pub struct AuthRequest {
   id: u64,
   response_type: ResponseType,
}

// Use `Query` extractor for query information (and destructure it within the signature).
// This handler gets called only if the request's query string contains `id` and `response_type` fields.
// The correct request for this handler would be `/index.html?id=64&response_type=Code"`.
async fn index(web::Query(info): web::Query<AuthRequest>) -> String {
    format!("Authorization request for client with id={} and type={:?}!", info.id, info.response_type)
}

fn main() {
    let app = App::new().service(
       web::resource("/index.html").route(web::get().to(index))); // <- use `Query` extractor
}

Implementations

impl<T> Query<T>

pub fn into_inner(self) -> T

Deconstruct to a inner value

pub fn from_query(query_str: &str) -> Result<Query<T>, QueryPayloadError> where
T: DeserializeOwned,

Get query parameters from the path