Module paperclip::v2::models

Models used by OpenAPI v2.

Structs

Api
OpenAPI v2 (swagger) spec generic over parameter and schema.

Coder
Represents the en/decoder for some MIME media range.

Coders
x-rust-coders global extension for custom encoders and decoders.

Contact
Contact object.

DefaultSchema
Default schema if your schema doesn’t have any custom fields.

DefaultSchemaRaw
Raw version of schema.

ExternalDocs
External Documentation object.

Header
Header object.

Info
Info object.

Items
Items object.

License
License object.

MediaRange
Wrapper for mime::MediaRange to support BTree{Set, Map}.

Operation
Operation object.

Parameter
Request parameter object.

PathItem
Path item object.

Reference
Reference object.

Response
Response object.

SecurityScheme
Security Scheme object.

Tag
Tag object.

Enums

CollectionFormat
Possible formats for array values in parameter.

DataType
Supported data types.

DataTypeFormat
Supported data type formats.

Either
Either from “either” crate. We can’t use that crate because we don’t want the enum to be tagged during de/serialization.

HttpMethod
The HTTP method used for an operation.

OperationProtocol
The protocol used for an operation.

ParameterIn
The location of the parameter.

Resolvable
Wrapper for schema. This uses Arc<RwLock<S>> for interior mutability and differentiates raw schema from resolved schema (i.e., the one where $ref references point to the actual schema).

SpecFormat
The format used by spec (JSON/YAML).

Version
OpenAPI version.

Statics

JSON_CODER
Default coder for JSON.

JSON_MIME
Media range for JSON.

YAML_CODER
Default coder for YAML.

YAML_MIME
Media range for YAML.

Type Definitions

DefaultApiRaw
OpenAPI v2 spec with defaults.

DefaultOperationRaw
Operation with default raw parameter and response.

DefaultParameterRaw
Parameter with the default raw schema.

DefaultPathItemRaw
Path item with default parameter and response.

DefaultResponseRaw
Response with the default raw schema.

ResolvableApi
OpenAPI v2 spec which can be traversed and resolved for codegen.

ResolvableOperation
Operation that can be traversed and resolved for codegen.

ResolvableParameter
Parameter that can be traversed and resolved for codegen.

ResolvablePathItem
Path item that can be traversed and resolved for codegen.

ResolvableResponse
Response that can be traversed and resolved for codegen.