Trait paperclip::v2::schema::Apiv2Schema

pub trait Apiv2Schema {
    const NAME: Option<&'static str>;

    fn raw_schema() -> DefaultSchemaRaw;

    fn schema_with_ref() -> DefaultSchemaRaw { ... }
}

Represents a OpenAPI v2 schema convertible. This is auto-implemented by api_v2_schema macro.

This is implemented for primitive types by default.

Associated Constants

const NAME: Option<&'static str>

Name of this schema. This is the object's name.

Loading content...
Required methods

fn raw_schema() -> DefaultSchemaRaw

Returns the raw schema for this object.

Loading content...
Provided methods

fn schema_with_ref() -> DefaultSchemaRaw

Returns the schema with a reference (if this is an object).

Here, we set the global reference to this object using its name, and we either remove the reference (remove_refs) or remove all properties other than the reference (retain_ref) based on where we're storing this object in the spec i.e., in an operation/response or in the map of definitions.

And we do that because at the time of this writing, statically collecting all models for all operations involved a lot of work, and so I went for runtime collection. Even though this happens at runtime, we're only doing this once at the start of the application, so it won't affect the incoming requests at all.

Loading content...